开发者模块-首页&APP维护页面-首页展示APP现在top5，维护页面筛选条件展示以及APP分类三级联动条件展示

+++ src/main/java/com/qf/entity/AppInfo.java
@@ -44,7 +44,7 @@ public class AppInfo  implements Serializable {
 	@Column(name = "flatform_id")
 	private Integer flatformId;
 
-	private Integer downloads;
+	private Long downloads;
 
 	@Column(name = "category_level1")
 	private Integer categoryLevel1;
